+++ src/test/java/ru/javawebinar/topjava/service/jdbc/JdbcMealServiceTest.java
@@ -1,0 +1,10 @@
+package ru.javawebinar.topjava.service.jdbc;
+
+import org.springframework.test.context.ActiveProfiles;
+import ru.javawebinar.topjava.service.AbstractMealServiceTest;
+
+import static ru.javawebinar.topjava.Profiles.JDBC;
+
+@ActiveProfiles(JDBC)
+public class JdbcMealServiceTest extends AbstractMealServiceTest {
+}
