+++ src/test/java/ru/javawebinar/topjava/service/jpa/JpaMealServiceTest.java
@@ -1,0 +1,10 @@
+package ru.javawebinar.topjava.service.jpa;
+
+import org.springframework.test.context.ActiveProfiles;
+import ru.javawebinar.topjava.service.AbstractMealServiceTest;
+
+import static ru.javawebinar.topjava.Profiles.JPA;
+
+@ActiveProfiles(JPA)
+public class JpaMealServiceTest extends AbstractMealServiceTest {
+}
